requirements.txt
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.11.1
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.5
dataclasses==0.8
decorator==5.1.1
entrypoints==0.4
Flask==2.0.3
idna==3.3
importlib-metadata==4.8.3
ipykernel==5.5.6
ipython==7.16.3
ipython-genutils==0.2.0
itsdangerous==2.0.1
jedi==0.17.2
Jinja2==3.0.3
jupyter-client==7.1.2
jupyter-core==4.9.2
MarkupSafe==2.0.1
MouseInfo==0.1.3
mysql-connector-python==8.0.29
nest-asyncio==1.5.5
numpy==1.19.5
pandas==1.1.5
parso==0.7.1
pickleshare==0.7.5
Pillow==8.4.0
prompt-toolkit==3.0.30
protobuf==3.19.4
PyAutoGUI==0.9.53
pycodestyle==2.8.0
PyGetWindow==0.0.9
Pygments==2.12.0
PyMsgBox==1.0.9
pyperclip==1.8.2
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
PyRect==0.2.0
PyScreeze==0.1.28
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.1
pywhatkit==5.4
pywin32==304
pyzmq==23.2.0
requests==2.27.1
six==1.16.0
soupsieve==2.3.2.post1
toml==0.10.2
tornado==6.1
traitlets==4.3.3
typing_extensions==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
Werkzeug==2.0.3
wikipedia==1.4.0
zipp==3.6.0
